Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@
 int main() {
     int m,n,o;
     scanf("%d %d %d",&m,&n,&o);
-    int sum = n+m+o;
+    float sum = n+m+o;
     printf("Average: %f",sum/3);
     return 0;
 }
